begin lifting the functionality of 'linear' into 'accelerate' terms
Work in progress, much left to do.

Some operations we won't be able to lift directly, e.g. those that use branching, but as far as possible try to avoid actually implementing anything: just tell accelerate how to use the implementations defined in linear.
