+++ linear-accelerate.cabal
@@ -4,14 +4,14 @@ version:       0.3
 license:       BSD3
 cabal-version: >= 1.8
 license-file:  LICENSE
-author:        Edward A. Kmett, Charles Durham
+author:        Edward A. Kmett, Charles Durham, Trevor L. McDonell
 maintainer:    Edward A. Kmett <ekmett@gmail.com>
 stability:     provisional
 homepage:      http://github.com/ekmett/linear-accelerate/
 bug-reports:   http://github.com/ekmett/linear-accelerate/issues
 copyright:     Copyright (C) 2014 Edward A. Kmett
-synopsis:      Instances to use linear vector spaces on accelerate backends
-description:   Instances to use linear vector spaces on accelerate backends
+synopsis:      Lifting linear vector spaces into Accelerate
+description:   Lifting linear vector spaces into Accelerate
 build-type:    Custom
 extra-source-files:
   .travis.yml
@@ -28,11 +28,19 @@ library
   build-depends:
     accelerate           >= 0.16,
     base                 >= 4.5   && < 5,
-    lens                 >= 4     && < 5,
     linear               >= 1.10  && < 2
 
   exposed-modules:
-    Linear.Accelerate
+    Data.Array.Accelerate.Linear
+    Data.Array.Accelerate.Linear.Metric
+    Data.Array.Accelerate.Linear.Plucker
+    Data.Array.Accelerate.Linear.Quaternion
+    Data.Array.Accelerate.Linear.V0
+    Data.Array.Accelerate.Linear.V1
+    Data.Array.Accelerate.Linear.V2
+    Data.Array.Accelerate.Linear.V3
+    Data.Array.Accelerate.Linear.V4
 
   ghc-options: -Wall -fwarn-tabs -O2
   hs-source-dirs: src
+
