+++ src/Data/Array/Accelerate/Linear/V1.hs
@@ -1,0 +1,63 @@
+{-# LANGUAGE ConstraintKinds       #-}
+{-# LANGUAGE FlexibleInstances     #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE ScopedTypeVariables   #-}
+{-# LANGUAGE TypeFamilies          #-}
+{-# LANGUAGE UndecidableInstances  #-}
+{-# OPTIONS_GHC -fno-warn-orphans #-}
+-----------------------------------------------------------------------------
+-- |
+-- Module      : Data.Array.Accelerate.Linear.V1
+-- Copyright   : 2014 Edward Kmett, Charles Durham,
+--               2015 Trevor L. McDonell
+-- License     : BSD-style (see the file LICENSE)
+--
+-- Maintainer  : Edward Kmett <ekmett@gmail.com>
+-- Stability   : experimental
+-- Portability : non-portable
+--
+-- 1-D Vectors
+----------------------------------------------------------------------------
+
+module Data.Array.Accelerate.Linear.V1 (
+
+  V1(..),
+
+) where
+
+import Data.Array.Accelerate
+import Data.Array.Accelerate.Smart
+import Data.Array.Accelerate.Product
+import Data.Array.Accelerate.Array.Sugar
+
+import Data.Array.Accelerate.Linear.Metric
+
+import Linear.V1                                ( V1(..) )
+
+
+-- Instances
+-- ---------
+
+instance Metric V1
+
+type instance EltRepr (V1 a) = EltRepr a
+
+instance Elt a => Elt (V1 a) where
+  eltType _ = eltType (undefined :: a)
+  toElt = V1 . toElt
+  fromElt (V1 a) = fromElt a
+
+instance cst a => IsProduct cst (V1 a) where
+  type ProdRepr (V1 a) = ((), a)
+  fromProd _ (V1 x) = ((), x)
+  toProd _ ((), x) = V1 x
+  prod _ _ = ProdRsnoc ProdRunit
+
+instance (Lift Exp a, Elt (Plain a)) => Lift Exp (V1 a) where
+  type Plain (V1 a) = V1 (Plain a)
+  lift (V1 x) = Exp . Tuple $ NilTup `SnocTup` lift x
+
+instance (Elt a, e ~ Exp a) => Unlift Exp (V1 e) where
+  unlift t = V1 $ Exp $ ZeroTupIdx `Prj` t
+
+
